added cashout and disp income calculations

+++ backend/ExpenseTrackerAPI/Models/Budget.cs
@@ -4,6 +4,6 @@ public class Budget
 {
 	public int Id { get; set; }
 	public int CashIn { get; set; }
-	public double CashOut { get; set; }
-	public double DisposableIncome { get; set; }
+	public int CashOut { get; set; }
+	public int DisposableIncome { get; set; }
 }
